[WIP] Implements WIN32

+++ turnip-engine/src/Graphics/Vulkan/PipelineVulkan.cpp
@@ -544,8 +544,8 @@ namespace tur
 			Scissor scissor;
 			{
 				scissor.offset = glm::vec2(0.f);
-				scissor.size.x = m_SwapchainExtent.width;
-				scissor.size.y = m_SwapchainExtent.height;
+				scissor.size.x = (uint32_t)m_SwapchainExtent.width;
+				scissor.size.y = (uint32_t)m_SwapchainExtent.height;
 			}
 			descriptor.scissors.push_back(scissor);
 		}
@@ -556,8 +556,8 @@ namespace tur
 		{
 			vk::Rect2D rect;
 			{
-				rect.offset = vk::Offset2D(scissor.offset.x, scissor.offset.y);
-				rect.extent = vk::Extent2D(scissor.size.x, scissor.size.y);
+				rect.offset = vk::Offset2D((int)scissor.offset.x, (int)scissor.offset.y);
+				rect.extent = vk::Extent2D((uint32_t)scissor.size.x, (uint32_t)scissor.size.y);
 			}
 
 			scissors.push_back(rect);
@@ -568,9 +568,9 @@ namespace tur
 		// Viewport State:
 		vk::PipelineViewportStateCreateInfo viewportState = { };
 		viewportState.flags = vk::PipelineViewportStateCreateFlags();
-		viewportState.viewportCount = viewports.size();
+		viewportState.viewportCount = (uint32_t)viewports.size();
 		viewportState.pViewports = viewports.data();
-		viewportState.scissorCount = scissors.size();
+		viewportState.scissorCount = (uint32_t)scissors.size();
 		viewportState.pScissors = scissors.data();
 		pipelineInfo.pViewportState = &viewportState;
 
@@ -596,7 +596,7 @@ namespace tur
 		shaderStages.push_back(fragmentShaderInfo);
 
 		// Pipeline Shader Stages:
-		pipelineInfo.stageCount = shaderStages.size();
+		pipelineInfo.stageCount = (uint32_t)shaderStages.size();
 		pipelineInfo.pStages = shaderStages.data();
 
 		// Multisampling
